scripts/requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
tabula-py==2.2.0
